# docs/openapi/alert-service.openapi.yaml
openapi: 3.0.3
info:
  title: Alert Service API
  version: FDP-26
...
servers:
  - url: http://localhost:8085
paths:
...
  /api/v1/fraud-cases:
    get:
      summary: List fraud cases
      description: >-
        Returns a paged fraud-case list. Size is bounded to 100. When filters are present, search is delegated to the
        fraud-case query service. This is a read-only endpoint and must not perform lifecycle mutation behavior.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/FraudCaseStatus"
        - name: assignee
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/FraudCasePriority"
        - name: riskLevel
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RiskLevel"
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: linkedAlertId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paged fraud cases returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseSummaryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create fraud case
      description: >-
        Creates a local audited fraud case from one or more alert ids. This endpoint is not idempotent unless a future
        contract explicitly states otherwise; clients must not blindly retry after an ambiguous network failure.
      security:
        - bearerAuth: []
        - demoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFraudCaseRequest"
      responses:
        "200":
          description: Fraud case created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/fraud-cases/{caseId}:
    get:
      summary: Get one fraud case
      description: Returns one fraud case for an authenticated analyst. Alert-service records a best-effort sensitive read-access audit event with backend-derived actor identity and bounded metadata only; audit persistence failure does not block the read response.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      responses:
        "200":
          description: Fraud case details returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update fraud case through regulated mutation coordinator
      description: Requires fraud-case:update and X-Idempotency-Key. Stores a regulated mutation command with canonical bounded intent fields for case id, backend-resolved actor, UPDATE_FRAUD_CASE, status, assignee hash, notes hash, tags hash, and payload hash. Matching idempotency replay returns the stored snapshot; conflicting key reuse returns 409. Non-terminal states return operation metadata and current_case_snapshot only; updated_case is present only when the mutation has committed. No raw fraud-case payloads, customer identifiers, or notes bodies are persisted in command intent.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
        - name: X-Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFraudCaseRequest"
      responses:
        "200":
          description: Fraud case update accepted or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateFraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/assign:
    post:
      summary: Assign fraud case
      description: >-
        Assigns a fraud case through the local audited lifecycle service. This endpoint is not idempotent unless a
        future contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignFraudCaseRequest"
      responses:
        "200":
          description: Fraud case assigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/notes:
    post:
      summary: Add fraud case note
      description: >-
        Appends a fraud-case note through the local audited lifecycle service. This endpoint is not idempotent unless
        a future contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFraudCaseNoteRequest"
      responses:
        "200":
          description: Fraud case note appended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseNoteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/decisions:
    post:
      summary: Add fraud case decision
      description: >-
        Appends a fraud-case decision through the local audited lifecycle service. This endpoint is not idempotent
        unless a future contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFraudCaseDecisionRequest"
      responses:
        "200":
          description: Fraud case decision appended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/transition:
    post:
      summary: Transition fraud case
      description: >-
        Transitions a fraud case through FraudCaseTransitionPolicy. This endpoint is not idempotent unless a future
        contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionFraudCaseRequest"
      responses:
        "200":
          description: Fraud case transitioned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/close:
    post:
      summary: Close fraud case
      description: >-
        Closes a fraud case through FraudCaseTransitionPolicy. This endpoint is not idempotent unless a future
        contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseFraudCaseRequest"
      responses:
        "200":
          description: Fraud case closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/reopen:
    post:
      summary: Reopen fraud case
      description: >-
        Reopens a fraud case through FraudCaseTransitionPolicy. This endpoint is not idempotent unless a future
        contract explicitly states otherwise.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReopenFraudCaseRequest"
      responses:
        "200":
          description: Fraud case reopened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/fraud-cases/{caseId}/audit:
    get:
      summary: Read fraud case audit trail
      description: Returns the append-only local fraud-case audit trail for users with fraud-case audit read authority.
      security:
        - bearerAuth: []
        - demoAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      responses:
        "200":
          description: Fraud case audit trail returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudCaseAuditResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    demoAuth:
      type: apiKey
      in: header
      name: X-Demo-User-Id
      description: Local/dev demo auth headers; roles are supplied through X-Demo-Roles.
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
...
    FraudCaseStatus:
      type: string
      enum: [OPEN, IN_REVIEW, ESCALATED, RESOLVED, REOPENED, CONFIRMED_FRAUD, FALSE_POSITIVE, CLOSED]
    FraudCasePriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    FraudCaseDecisionType:
      type: string
      enum: [FRAUD_CONFIRMED, FALSE_POSITIVE, NEEDS_MORE_INFO, ESCALATE, NO_ACTION]
    FraudCaseAuditAction:
      type: string
      enum: [CASE_CREATED, CASE_ASSIGNED, CASE_REASSIGNED, NOTE_ADDED, DECISION_ADDED, STATUS_CHANGED, CASE_CLOSED, CASE_REOPENED]
    CreateFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [alertIds, priority, riskLevel, actorId]
      properties:
        alertIds:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
            maxLength: 128
        priority:
          $ref: "#/components/schemas/FraudCasePriority"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        reason:
          type: string
          nullable: true
          maxLength: 500
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    AssignFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [assignedInvestigatorId, actorId]
      properties:
        assignedInvestigatorId:
          type: string
          minLength: 1
          maxLength: 64
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    AddFraudCaseNoteRequest:
      type: object
      additionalProperties: false
      required: [body, actorId]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
        internalOnly:
          type: boolean
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    AddFraudCaseDecisionRequest:
      type: object
      additionalProperties: false
      required: [decisionType, summary, actorId]
      properties:
        decisionType:
          $ref: "#/components/schemas/FraudCaseDecisionType"
        summary:
          type: string
          minLength: 1
          maxLength: 2000
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    TransitionFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [targetStatus, actorId]
      properties:
        targetStatus:
          $ref: "#/components/schemas/FraudCaseStatus"
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    CloseFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [closureReason, actorId]
      properties:
        closureReason:
          type: string
          minLength: 1
          maxLength: 500
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    ReopenFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [reason, actorId]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
        actorId:
          type: string
          minLength: 1
          maxLength: 64
    FraudCaseSummaryPage:
      type: object
      additionalProperties: false
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/FraudCaseSummaryResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
    FraudCaseSummaryResponse:
      type: object
      additionalProperties: false
      properties:
        caseId:
          type: string
        caseNumber:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/FraudCaseStatus"
        priority:
          $ref: "#/components/schemas/FraudCasePriority"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        assignedInvestigatorId:
          type: string
          nullable: true
        linkedAlertIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FraudCaseResponse:
      type: object
      additionalProperties: false
      properties:
        caseId:
          type: string
        caseNumber:
          type: string
          nullable: true
        customerId:
          type: string
          nullable: true
        suspicionType:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/FraudCaseStatus"
        priority:
          $ref: "#/components/schemas/FraudCasePriority"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        linkedAlertIds:
          type: array
          items:
            type: string
        assignedInvestigatorId:
          type: string
          nullable: true
        createdBy:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        thresholdPln:
          type: number
          nullable: true
        totalAmountPln:
          type: number
          nullable: true
        aggregationWindow:
          type: string
          nullable: true
        firstTransactionAt:
          type: string
          format: date-time
          nullable: true
        lastTransactionAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        analystId:
          type: string
          nullable: true
        decisionReason:
          type: string
          nullable: true
        decisionTags:
          type: array
          items:
            type: string
        decidedAt:
          type: string
          format: date-time
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        closureReason:
          type: string
          nullable: true
        version:
          type: integer
          format: int64
          nullable: true
        transactionIds:
          type: array
          items:
            type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/FraudCaseTransactionResponse"
    FraudCaseTransactionResponse:
      type: object
      additionalProperties: false
      properties:
        transactionId:
          type: string
        correlationId:
          type: string
          nullable: true
        transactionTimestamp:
          type: string
          format: date-time
          nullable: true
        transactionAmount:
          $ref: "#/components/schemas/GenericObject"
        amountPln:
          type: number
          nullable: true
        fraudScore:
          type: number
          nullable: true
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
    FraudCaseNoteResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        caseId:
          type: string
        body:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        internalOnly:
          type: boolean
    FraudCaseDecisionResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        caseId:
          type: string
        decisionType:
          $ref: "#/components/schemas/FraudCaseDecisionType"
        summary:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    FraudCaseAuditResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        caseId:
          type: string
        action:
          $ref: "#/components/schemas/FraudCaseAuditAction"
        actorId:
          type: string
        occurredAt:
          type: string
          format: date-time
        previousStatus:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/FraudCaseStatus"
        newStatus:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/FraudCaseStatus"
        details:
          type: object
          additionalProperties:
            type: string
    UpdateFraudCaseRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FraudCaseStatus"
        analystId:
          type: string
          nullable: true
          maxLength: 120
          description: Optional frontend-provided hint only; backend principal remains authoritative.
        decisionReason:
          type: string
          nullable: true
          maxLength: 1000
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 80
    UpdateFraudCaseResponse:
      type: object
      additionalProperties: false
      required: [operation_status, case_id]
      properties:
        operation_status:
          type: string
          enum: [REJECTED_BEFORE_MUTATION, IN_PROGRESS, RECOVERY_REQUIRED, COMMIT_UNKNOWN, EVIDENCE_PREPARING, EVIDENCE_PREPARED, FINALIZING, FINALIZED_VISIBLE, FINALIZED_EVIDENCE_PENDING_EXTERNAL, FINALIZED_EVIDENCE_CONFIRMED, REJECTED_EVIDENCE_UNAVAILABLE, FAILED_BUSINESS_VALIDATION, FINALIZE_RECOVERY_REQUIRED, COMMITTED_EVIDENCE_PENDING, COMMITTED_EVIDENCE_CONFIRMED, COMMITTED_EVIDENCE_INCOMPLETE]
          description: Mandatory operation state. IN_PROGRESS, RECOVERY_REQUIRED, and COMMIT_UNKNOWN must not be interpreted as a committed fraud-case update.
        command_id:
          type: string
          nullable: true
        idempotency_key_hash:
          type: string
          nullable: true
        case_id:
          type: string
        current_case_snapshot:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/GenericObject"
        updated_case:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/GenericObject"
          description: Present only for committed/evidence-pending committed states. Absent for IN_PROGRESS, RECOVERY_REQUIRED, COMMIT_UNKNOWN, and rejected states.
        recovery_required_reason:
          type: string
          nullable: true
...
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - details
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Authentication is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Authenticated actor lacks the required authority.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...